requirements.txt
streamlit
sentence_transformers
chromadb
google-generativeai
ipython
pdfplumber
openai
tiktoken
nltk